removed authentication flow from frontend
